Fix UI Errors for Card
The card wasnt showing for some reason, and somehow the playmats were starting as an existing color. Which may confuse players who want to play that color.

Also added an upside-down number to the card for aesthetic purposes. The number is no longer random on start-up instead it defaults to 1.